JS: merge KotlinJavascriptSerializer into KotlinJavascriptSerializationUtil
