+++ js/js.serializer/src/org/jetbrains/kotlin/serialization/js/ClassSerializationUtil.kt
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.jetbrains.kotlin.utils.serializer
+package org.jetbrains.kotlin.serialization.js
 
 import org.jetbrains.kotlin.descriptors.ClassDescriptor
-import org.jetbrains.kotlin.serialization.ProtoBuf
-import org.jetbrains.kotlin.serialization.DescriptorSerializer
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
+import org.jetbrains.kotlin.serialization.DescriptorSerializer
+import org.jetbrains.kotlin.serialization.ProtoBuf
 
 public object ClassSerializationUtil {
     public trait Sink {
