+++ js/js.serializer/src/org/jetbrains/kotlin/serialization/js/KotlinJavascriptSerializationUtil.kt
@@ -17,12 +17,23 @@
 package org.jetbrains.kotlin.serialization.js
 
 import com.google.protobuf.ByteString
+import org.jetbrains.kotlin.builtins.BuiltInsSerializationUtil
 import org.jetbrains.kotlin.builtins.BuiltinsPackageFragment
+import org.jetbrains.kotlin.descriptors.ClassDescriptor
+import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.descriptors.PackageFragmentProvider
 import org.jetbrains.kotlin.name.FqName
+import org.jetbrains.kotlin.resolve.DescriptorUtils
+import org.jetbrains.kotlin.resolve.descriptorUtil.classId
+import org.jetbrains.kotlin.resolve.scopes.DescriptorKindFilter
+import org.jetbrains.kotlin.serialization.DescriptorSerializer
+import org.jetbrains.kotlin.serialization.ProtoBuf
+import org.jetbrains.kotlin.serialization.SerializationUtil
+import org.jetbrains.kotlin.serialization.builtins.BuiltInsSerializerExtension
 import org.jetbrains.kotlin.serialization.deserialization.FlexibleTypeCapabilitiesDeserializer
 import org.jetbrains.kotlin.storage.LockBasedStorageManager
+import org.jetbrains.kotlin.utils.KotlinJavascriptMetadataUtils
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.util.zip.GZIPInputStream
@@ -71,6 +82,54 @@ public object KotlinJavascriptSerializationUtil {
         gzipOutputStream.close()
 
         return byteStream.toByteArray()
+    }
+
+    public fun metadataAsString(moduleName: String, moduleDescriptor: ModuleDescriptor): String {
+        val contentMap = hashMapOf<String, ByteArray>()
+
+        DescriptorUtils.getPackagesFqNames(moduleDescriptor).forEach {
+            fqName ->
+            serializePackage(moduleDescriptor, fqName) {
+                fileName, stream ->
+                contentMap[fileName] = stream.toByteArray()
+            }
+        }
+
+        val content = KotlinJavascriptSerializationUtil.contentMapToByteArray(contentMap)
+        return KotlinJavascriptMetadataUtils.formatMetadataAsString(moduleName, content)
+    }
+
+    fun serializePackage(module: ModuleDescriptor, fqName: FqName, writeFun: (String, ByteArrayOutputStream) -> Unit) {
+        val packageView = module.getPackage(fqName) ?: error("No package resolved in $module")
+
+        val skip: (DeclarationDescriptor) -> Boolean = { DescriptorUtils.getContainingModule(it) != module }
+
+        val serializer = DescriptorSerializer.createTopLevel(BuiltInsSerializerExtension)
+
+        val classifierDescriptors = DescriptorSerializer.sort(packageView.getMemberScope().getDescriptors(DescriptorKindFilter.CLASSIFIERS))
+
+        ClassSerializationUtil.serializeClasses(classifierDescriptors, serializer, object : ClassSerializationUtil.Sink {
+            override fun writeClass(classDescriptor: ClassDescriptor, classProto: ProtoBuf.Class) {
+                val stream = ByteArrayOutputStream()
+                classProto.writeTo(stream)
+                writeFun(getFileName(classDescriptor), stream)
+            }
+        }, skip)
+
+        val packageStream = ByteArrayOutputStream()
+        val fragments = module.getPackageFragmentProvider().getPackageFragments(fqName)
+        val packageProto = serializer.packageProto(fragments, skip).build() ?: error("Package fragments not serialized: $fragments")
+        packageProto.writeTo(packageStream)
+        writeFun(BuiltInsSerializationUtil.getPackageFilePath(fqName), packageStream)
+
+        val nameStream = ByteArrayOutputStream()
+        val strings = serializer.getStringTable()
+        SerializationUtil.serializeStringTable(nameStream, strings.serializeSimpleNames(), strings.serializeQualifiedNames())
+        writeFun(BuiltInsSerializationUtil.getStringTableFilePath(fqName), nameStream)
+    }
+
+    fun getFileName(classDescriptor: ClassDescriptor): String {
+        return BuiltInsSerializationUtil.getClassMetadataPath(classDescriptor.classId)
     }
 
     private fun getPackages(contentMap: Map<String, ByteArray>): List<String> =
